refactor: URL 필터링 로직을 useURLFilters 훅으로 추출
