+++ apps/admin/src/shared/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useURLFilters } from './useURLFilters';
